Fix parsing of the `mesh.msh` output file
Instead of adding the `mesh.msh` output file to `retrieve_list` we add
it to the `retrieve_temporary_list`. Since the parser will wrap the file
in its entirety in a `SinglefileNode` and report it as an output node,
there is no need to also store it in the `retrieved` output node, which
will essentially duplicate the file.

By retrieving the file in the temporary folder, it will be available for
the parser to create the output node from. The parser is adapted to read
the output file from the `retrieved_temporary_folder`, whose filepath is
passed in the `kwargs` of the `Parser.parse` call.